% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allocte_trials.R
\name{allocate_trials}
\alias{allocate_trials}
\title{Title}
\usage{
allocate_trials(
  df,
  index_pairs,
  chunks_per_trial = 1,
  reset_to_1 = TRUE,
  fill_StudioEventData = TRUE
)
}
\arguments{
\item{df}{df}

\item{index_pairs}{index_pairs}

\item{chunks_per_trial}{chunks_per_trial}

\item{reset_to_1}{reset_to_1}

\item{fill_StudioEventData}{fill_StudioEventData}
}
\value{
None
}
\description{
Title
}
