% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_fi.R
\name{create_fi}
\alias{create_fi}
\title{Create a Fixation Index Column based on gaze type and gaze type index}
\usage{
create_fi(df, type_col, type_index_col)
}
\arguments{
\item{df}{raw dataframe}

\item{type_col}{the column name of the column that specifies gaze type (fixation, saccaded, etc)}

\item{type_index_col}{the column name containing indexes for the specific type_col}
}
\value{
df including a FixationIndex column
}
\description{
eyewit requires a FixationIndex column (i.e., \code{fi}) in which fixation indexes (and only
fixation indexes) are in a sequential order (i.e., 1, 2, 3, NA, NA, 4, 5, ...).
}
\details{
Tobii Studio had a FixationIndex column on which eyewit heavily relies on. Since Tobii ProLab
put that information in two columns: \verb{Eye movement type} and \verb{Eye movement type index}, we need
to filter \verb{Eye movement type} for "Fixation" only to create a new column \code{FixationIndex} with
the corresponding values from \verb{Eye movement type index}.
}
