% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fi2rn.R
\name{fi2rn}
\alias{fi2rn}
\title{Map FixationIndexes to RowNumbers (fi2rn)}
\usage{
fi2rn(fi_col)
}
\arguments{
\item{fi_col}{A column vector containing fixation indexes}
}
\value{
A named list
}
\description{
Given a vector containing fixation indexes, this function returns a named list containing
corresponding row names to the given fixation indexes. Since fixation indexes span a range
for each index, the function uses the two groups: \code{fistart} and \code{fiend} to separate them.
That means, \verb{$fistart} denotes all row names where a fixation index is starting, wheras \verb{$fiend}
denotes all row names where the fixation index ends.
Since the function is used in several places within the package, it gets
\strong{executed during the preflight automatically}. That means you don’t need to call this
function on your own. To access the content, you can call the \strong{variable}:
\code{.eyewit_util$fi2rn} (see examples).
}
\examples{
\dontrun{
# Assume you have fixation indexes in a df with a column named FixationIndex

# returns all $fistart and $fiend row names and store it (this is what happens in preflight)
fi2rn <- fi2rn(df$FixationIndex)

# now call the variable because the function call is expensive

# get the first row number of the first fixation index
fi2rn(df$FixationIndex)$fistart[1]

# get the last row number of the first fixation index
fi2rn(df$FixationIndex)$fiend[1]

# get the last row number of the 42. fixation index
fi2rn(df$FixationIndex)$fiend[42]
}
}
\seealso{
Other preflight functions: 
\code{\link{is_aoilist_intersecting}()},
\code{\link{is_intersecting}()}
}
\concept{preflight functions}
