% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fi_summary.R
\name{fi_summary}
\alias{fi_summary}
\title{Fixation Index Summary}
\usage{
fi_summary(df, aoisets = NA, trialrange = NA, show_non_hn_labels = FALSE)
}
\arguments{
\item{df}{A dataframe containing columns created by \link{get_aois}.}

\item{aoisets}{\emph{(Optional)} A list of all AOI sets, or a single AOI definition.}

\item{trialrange}{\emph{(Optional)} A trial range (trlr) list.}

\item{show_non_hn_labels}{\emph{(Optional)} TRUE or FALSE (default).}
}
\value{
Returns a dataframe (tibble) providing AOI information for all fixation indexes
}
\description{
Given a df and, optionally, a set of AOI definitions, this function returns a df providing a
summary for all fixation indexes. Specifically, for every fixation index you will get the row
name range, timestamps, fixation durations, and the information which AOI was hit at the given
fixation index. See Arguments Details section further down for detailed explanation.
}
\section{Argument Details}{
Depending on the arguments you provide, the returned df varies as such:
\subsection{df (mandatory)}{

If you only provide a df, this function is identical to \link{fi2rn}. It is thus cheaper to call
\link{fi2rn}, if you only need to know the row range for a given fixation index.
}

\subsection{df, aoisets}{

Returns a tibble respecting all AOI definitions (or just a list depending on your input)

\if{html}{\out{<div class="sourceCode">}}\preformatted{tribble(
  ~FI, ~FIrnS, ~FIrnE, ~AOIColA,             ~AOIColN,
  1,   15,        30,      c("left"),            c("onscreen"),
  2,   45,        70,      c("right"),           c("onscreen"),
  3,   80,        99,      c("left", "outside"), c("onscreen"),
)
}\if{html}{\out{</div>}}
}

\subsection{df, aoisets, show_non_hn_labels = TRUE}{

Returns a tibble respecting all AOI definitions, and includes a clean-column
without non-hitname labels. Non-hitname labels are defined in the \code{interface.R} file. Look for:
\code{no_evaluation_label}, \code{missing_coordinate_label}, and \code{outside_aoi_label}.
Compare this tribble with the previous one:

\if{html}{\out{<div class="sourceCode">}}\preformatted{tribble(
  ~FI, ~FIrnS, ~FIrnE, ~AOIColA,             ~AOIColAClean, ~AOIColN,      ~AOIColNClean,
  1,   15,        30,      c("left"),            c("left"),     c("onscreen"), NA,
  2,   45,        70,      c("right"),           c("right"),    c("onscreen"), NA,
  3,   80,        99,      c("left", "outside"), c("left"),     c("onscreen"), NA,
)
}\if{html}{\out{</div>}}
}

\subsection{df, aoisets, trial range}{

If you provide a trial range, the returned df will be trimmed down to respect the row ranges. It
will use have an additional column \code{FGapDurTrlEnd} which is similar to \code{FGapDur}; yet, this
column will respect the trial range end in the calculation for the duration. For example,
if the fixation index 420 is the last one in trail 7, \code{FGapDur} will show the time difference
from 420 to 421. In contrast, \code{FGapDurTrlEnd} will show the time difference from 420 to the
trialrng end time (e.g., "MovieEnd"). Thus, \code{FGapDurTrlEnd} is naturally smaller than \code{FGapDur}.
}
}

\examples{
\dontrun{
# Returns a df respecting the aoi definitions from the interface
fi_summary(df, aoisets)
}
}
