% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_aois.R
\name{get_aois}
\alias{get_aois}
\title{Title}
\usage{
get_aois(x, y, aoi_collection, scope = NA)
}
\arguments{
\item{x}{x}

\item{y}{y}

\item{aoi_collection}{aoi_collection}

\item{scope}{scope}
}
\value{
None
}
\description{
Title
}
