% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fixationindex_pairs.R
\name{get_fixationindex_pairs}
\alias{get_fixationindex_pairs}
\title{Get FixationIndex pairs}
\usage{
get_fixationindex_pairs(fi_col)
}
\arguments{
\item{fi_col}{A column vector containing FixationIndexes}
}
\value{
A named list which contains rownumbers for a given FixationIndex
}
\description{
$fistart is the first and $fiend the last value for a given single FixationIndex
}
\examples{
# Todo get_fixationindex_pairs(df$FixationIndex)$fistart[110] returns a rn of 6835
}
