% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gazeshift_latency.R
\name{get_gazeshift_latency}
\alias{get_gazeshift_latency}
\title{Title}
\usage{
get_gazeshift_latency(df, aoisets)
}
\arguments{
\item{df}{A dataframe}

\item{aoisets}{a AOI set}
}
\value{
None
}
\description{
https://github.com/Kalaschnik/media/blob/main/get_gazeshift_latency.png
}
