% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lookaway_scope_end.R
\name{get_lookaway_scope_end_deprecated}
\alias{get_lookaway_scope_end_deprecated}
\title{Title}
\usage{
get_lookaway_scope_end_deprecated(df, aoi_collection, scope, lookaway_stop)
}
\arguments{
\item{df}{df}

\item{scope}{scope}

\item{lookaway_stop}{lookaway_stop}
}
\value{
None
}
\description{
Title
}
