% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_looks.R
\name{get_looks}
\alias{get_looks}
\title{Title}
\usage{
get_looks(
  df,
  aoi_collection,
  scope = NA,
  intra_scope_window = c("start", "end"),
  lookaway_stop = NA,
  omit_first_overflow_fi = FALSE,
  first_look_emergency_cutoff = NA,
  intra_scope_cut = TRUE
)
}
\arguments{
\item{df}{df}

\item{aoi_collection}{aoi_collection}

\item{scope}{scope}

\item{intra_scope_window}{intra_scope_window}

\item{lookaway_stop}{lookaway_stop}

\item{omit_first_overflow_fi}{omit_first_overflow_fi}

\item{first_look_emergency_cutoff}{first_look_emergency_cutoff}

\item{intra_scope_cut}{intra_scope_cut}
}
\value{
None
}
\description{
Title
}
