% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_non_fixation_data.R
\name{get_non_fixation_data_deprecated}
\alias{get_non_fixation_data_deprecated}
\title{Title}
\usage{
get_non_fixation_data_deprecated(df, scope = NA)
}
\arguments{
\item{df}{df}

\item{scope}{scope}
}
\value{
None
}
\description{
Title
}
