% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_preflook_pos.R
\name{get_preflook_pos}
\alias{get_preflook_pos}
\title{Title}
\usage{
get_preflook_pos(preflook_string_vector, familiar_object_vector)
}
\arguments{
\item{preflook_string_vector}{preflook_string_vector}

\item{familiar_object_vector}{familiar_object_vector}
}
\value{
None
}
\description{
Title
}
