% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_starend_pos.R
\name{get_start_end_pos}
\alias{get_start_end_pos}
\title{Get Start & End Positions}
\usage{
get_start_end_pos(
  df,
  regex_pattern,
  event_start = "MovieStart",
  event_end = "MovieEnd"
)
}
\arguments{
\item{df}{A df which includes Tobii/E-Prime start and stop Markers.}

\item{regex_pattern}{A regular expression.}

\item{event_start}{The name of the start marker (e.g. MovieStart).}

\item{event_end}{The name of the end marker (MovieEnd)}
}
\value{
Returns a named list of start and stop positions for a given regexp.
}
\description{
Return a named list of start and stop positions for a given regular expression based on
event column, which contains start and end markers
}
\details{
Note that the return pairs will be NOT BE shifted with +1 for Start Events or -1 for End Events
this is due to Tobii’s bad behavior to sometimes stack events https://pasteboard.co/J24dbgx.png
}
\examples{
# get_StartEnd_list(coi_df, ".*Familiar.*", "MovieStart", "MovieEnd")
}
