% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_hitname_in_range.R
\name{is_hitname_in_range}
\alias{is_hitname_in_range}
\title{Is hitname in range}
\usage{
is_hitname_in_range(vec, hitname, fi_start, fi_end)
}
\arguments{
\item{vec}{vec}

\item{hitname}{hitname}

\item{fi_start}{fi_start}

\item{fi_end}{fi_end}
}
\value{
None
}
\description{
this function accepts a column containing hitnames, a target hitname to look for, and a
starting and ending fixation index
}
