% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_startend_chunks.R
\name{merge_startend_chunks}
\alias{merge_startend_chunks}
\title{Title}
\usage{
merge_startend_chunks(se_lists, shared_key = "start")
}
\arguments{
\item{se_lists}{se_lists}

\item{shared_key}{shared_key}
}
\value{
None
}
\description{
Title
}
