% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timebinning.R
\name{timebinning}
\alias{timebinning}
\title{Time Binning}
\usage{
timebinning(df, df_subject, scope, bin_width = 500)
}
\arguments{
\item{df}{A dataframe}

\item{df_subject}{A dedicated df for an individual subject, containing a "Condition" column}

\item{scope}{A starend list to determine the length in ms for each trial}

\item{bin_width}{The size of the bin}
}
\value{
A df for an individual subject
}
\description{
Returns binned looking times for an individual subject
}
\examples{
\dontrun{
}
}
