% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/value_parser_by_key.R
\name{value_parser_by_key}
\alias{value_parser_by_key}
\title{Title}
\usage{
value_parser_by_key(
  keys,
  delimited_string,
  delimiter = "_",
  drop_nth_delimiter,
  trim_right
)
}
\arguments{
\item{keys}{keys}

\item{delimited_string}{delimited_string}

\item{delimiter}{delimiter}

\item{drop_nth_delimiter}{drop_nth_delimiter}

\item{trim_right}{trim_right}
}
\value{
None
}
\description{
Title
}
\examples{
## usage:
# keys <- c("age", "sex", "location", "language")
# delimited_string <- "36_f_leipzig_german.mp4"
# delimiter = "_"
# drop_nth_delimiter = 2
# trim_right = 4
# location <- value_parser_by_key(keys, delimited_string, trim_right = 4)$language # german
# location <- value_parser_by_key(c("age"), delimited_string, drop_nth_delimiter = c(1,2,3))$age
# ...36fleipziggerman.mp4
}
