% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vendor_check.R
\name{vendor_check}
\alias{vendor_check}
\title{vendor_check}
\usage{
vendor_check(df)
}
\arguments{
\item{df}{Your raw dataframe}
}
\value{
Returns a \code{vendor_lookup} key as string
}
\description{
vendor_check
}
