% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fi_summary.R
\name{fi_summary}
\alias{fi_summary}
\title{Fixation Index Summary}
\usage{
fi_summary(
  df,
  aoisets = NA,
  scope = NA,
  show_non_hn_labels = FALSE,
  na_trailing_gap_duration = TRUE
)
}
\arguments{
\item{df}{A dataframe containing columns created by \link{get_aois}.}

\item{scope}{\emph{\link{Optional}} A trial scope list.}

\item{show_non_hn_labels}{\emph{\link{Optional}} TRUE or FALSE (default).}

\item{na_trailing_gap_duration}{\emph{\link{Optional}} TRUE (default) or FALSE.}

\item{aoi_sets}{\emph{\link{Optional}} A list of AOI sets.}
}
\value{
Returns a dataframe (tibble) providing AOI information for all fixation indexes
}
\description{
Given a df and, optionally, a set of AOI definitions, this function returns a df providing a
summary for all fixation indexes. Specifically, for every fixation index you will get the row
name range, timestamps, fixation durations, and the information which AOI was hit at the given
fixation index. See Arguments Details section further down for detailed explanation.
}
\section{Argument Details}{
Depending on the arguments you provide, the returned df varies as such:
\subsection{df (mandatory)}{

If you only provide a df, this function is identical to \link{fi2rn}. It is thus cheaper to call
\link{fi2rn}, if you only need to know the row range for a given fixation index.
}

\subsection{df, aoisets}{

Returns a tibble respecting all AOI definitions\preformatted{tribble(
  ~FI, ~RowStart, ~RowEnd, ~AOIColA,             ~AOIColN,
  1,   15,        30,      c("left"),            c("onscreen"),
  2,   45,        70,      c("right"),           c("onscreen"),
  3,   80,        99,      c("left", "outside"), c("onscreen"),
)
}
}

\subsection{df, aoisets, show_non_hn_labels = TRUE}{

Returns a tibble respecting all AOI definitions, and includes a clean-column
without non-hitname labels. Non-hitname labels are defined in the \code{interface.R} file. Look for:
\code{no_evaluation_label}, \code{missing_coordinate_label}, and \code{outside_aoi_label}.
Compare this tribble with the previous one:\preformatted{tribble(
  ~FI, ~RowStart, ~RowEnd, ~AOIColA,             ~AOIColAClean, ~AOIColN,      ~AOIColNClean,
  1,   15,        30,      c("left"),            c("left"),     c("onscreen"), NA,
  2,   45,        70,      c("right"),           c("right"),    c("onscreen"), NA,
  3,   80,        99,      c("left", "outside"), c("left"),     c("onscreen"), NA,
)
}
}

\subsection{df, aoisets, trial scope}{

If you provide a trial scopes, the returned df will be trimmed down to respect the row ranges.
\emph{NB}, when using trial scopes, the last value for every scope end at \strong{FixationGapDuration} will
show NA. This behavior is on purpose. Because if a trial block ends, usually another media
elements starts and there is a longer inactive scence, you want to avoid having these larger
values in your data. For example, the duration between the last fixation index of scene A and
first fixation index of scene B is naturally higher, and should not be evaluated.
Since a scene/trial was ending anyways, it is not a good proxy for detecting a look-away
during a trial. You can bypass this option by setting \code{na_trailing_gap_duration} to FALSE.
}
}

\examples{
\dontrun{
# Returns a df respecting the aoi definitions from the interface
fi_summary(df, aoisets)
}
}
