% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_aoilist_intersecting.R
\name{is_aoilist_intersecting}
\alias{is_aoilist_intersecting}
\title{Check if AOIs intersect}
\usage{
is_aoilist_intersecting(aoisets, verbose = FALSE)
}
\arguments{
\item{aoisets}{An AOI object containing arbitrary AOI lists containing arbitrary AOIs as defined in interface.R}

\item{verbose}{Enable verbose output (defaults to FALSE)}
}
\value{
Returns a logical scalar
}
\description{
The first input is the aoisets list as defined in the interface.R file. The function permutates
over all AOIs and checks if they are intersecting
}
\details{
This function uses \link{is_intersecting}
}
\examples{
#' is_aoilist_intersecting(
  list(x_topleft = 10, y_topleft = 10, x_bottomright = 20, y_bottomright = 20),
  list(x_topleft = 15, y_topleft = 15, x_bottomright = 25, y_bottomright = 25)
) # Rectangles are clearly intersecting (⧉) → returns TRUE

#' is_aoilist_intersecting(
  list(x_topleft = 10, y_topleft = 10, x_bottomright = 20, y_bottomright = 20),
  list(x_topleft = 15, y_topleft = 15, x_bottomright = 25, y_bottomright = 25)
) # Rectangles are clearly intersecting (⧉) → returns TRUE

}
\seealso{
Other preflight functions: 
\code{\link{is_intersecting}()}
}
\concept{preflight functions}
